+++ src/redturtle/volto/restapi/deserializer/blocks.py
@@ -32,7 +32,7 @@ class GenericResolveUIDDeserializer(object):
 
     def fix_urls_in_block(self, block):
         if isinstance(block, str):
-            return path2uid(context=self.context, link=block)
+            return self.get_uid_from_path(link=block)
         if block.get("@type", "") in EXCLUDE_TYPES:
             return block
         if "UID" in block.keys():
@@ -44,7 +44,7 @@ class GenericResolveUIDDeserializer(object):
             if key in EXCLUDE_KEYS:
                 continue
             if isinstance(val, str):
-                block[key] = path2uid(context=self.context, link=val)
+                block[key] = self.get_uid_from_path(link=val)
             elif isinstance(val, list):
                 block[key] = [self.fix_urls_in_block(x) for x in val]
             elif isinstance(val, dict):
@@ -52,21 +52,30 @@ class GenericResolveUIDDeserializer(object):
                     entity_map = val.get("entityMap", {})
                     for entity_map in entity_map.values():
                         url = entity_map["data"].get("url", "").strip("/")
-                        entity_map["data"]["url"] = path2uid(
-                            context=self.context, link=url
-                        )
+                        entity_map["data"]["url"] = self.get_uid_from_path(link=url)
                 else:
                     block[key] = self.fix_urls_in_block(block=val)
         return block
+
+    def get_uid_from_path(self, link):
+        """get_uid_from_path.
+
+        :param link:
+        """
+        try:
+            return path2uid(context=self.context, link=link)
+        except IndexError:
+            # the value (link) is not a valid path
+            return link
 
 
 @implementer(IBlockFieldDeserializationTransformer)
 @adapter(IBlocks, IRedturtleVoltoLayer)
 class GenericResolveUIDDeserializerContents(GenericResolveUIDDeserializer):
-    """ Deserializer for content-types that implements IBlocks behavior """
+    """Deserializer for content-types that implements IBlocks behavior"""
 
 
 @implementer(IBlockFieldDeserializationTransformer)
 @adapter(IPloneSiteRoot, IRedturtleVoltoLayer)
 class GenericResolveUIDDeserializerRoot(GenericResolveUIDDeserializer):
-    """ Deserializer for site-root """
+    """Deserializer for site-root"""